Extract functions related to swapping configuration in their own module
