+++ lib/spack/spack/bootstrap/__init__.py
@@ -1,3 +1,12 @@
+# Copyright 2013-2022 Lawrence Livermore National Security, LLC and other
+# Spack Project Developers. See the top-level COPYRIGHT file for details.
+#
+# SPDX-License-Identifier: (Apache-2.0 OR MIT)
 """Function and classes needed to bootstrap Spack itself."""
 
 from .bootstrap import *  # noqa: F403,F401
+from .config import (  # noqa: F401
+    ensure_bootstrap_configuration,
+    spack_python_interpreter,
+    store_path,
+)
