+++ lib/spack/spack/bootstrap/bootstrap.py
@@ -4,7 +4,6 @@
 # SPDX-License-Identifier: (Apache-2.0 OR MIT)
 from __future__ import print_function
 
-import contextlib
 import copy
 import fnmatch
 import functools
@@ -41,6 +40,12 @@ import spack.util.path
 import spack.util.spack_yaml
 import spack.util.url
 import spack.version
+
+from .config import (
+    ensure_bootstrap_configuration,
+    spack_python_interpreter,
+    spec_for_current_python,
+)
 
 #: Name of the file containing metadata about the bootstrapping source
 METADATA_YAML_FILENAME = "metadata.yaml"
@@ -466,36 +471,6 @@ def source_is_enabled_or_raise(conf):
         raise ValueError("source is not trusted")
 
 
-def spec_for_current_python():
-    """For bootstrapping purposes we are just interested in the Python
-    minor version (all patches are ABI compatible with the same minor).
-
-    See:
-      https://www.python.org/dev/peps/pep-0513/
-      https://stackoverflow.com/a/35801395/771663
-    """
-    version_str = ".".join(str(x) for x in sys.version_info[:2])
-    return "python@{0}".format(version_str)
-
-
-@contextlib.contextmanager
-def spack_python_interpreter():
-    """Override the current configuration to set the interpreter under
-    which Spack is currently running as the only Python external spec
-    available.
-    """
-    python_prefix = sys.exec_prefix
-    external_python = spec_for_current_python()
-
-    entry = {
-        "buildable": False,
-        "externals": [{"prefix": python_prefix, "spec": str(external_python)}],
-    }
-
-    with spack.config.override("packages:python::", entry):
-        yield
-
-
 def ensure_module_importable_or_raise(module, abstract_spec=None):
     """Make the requested module available for import, or raise.
 
@@ -617,31 +592,6 @@ def _python_import(module):
     return True
 
 
-def _bootstrap_config_scopes():
-    tty.debug("[BOOTSTRAP CONFIG SCOPE] name=_builtin")
-    config_scopes = [spack.config.InternalConfigScope("_builtin", spack.config.config_defaults)]
-    configuration_paths = (spack.config.configuration_defaults_path, ("bootstrap", _config_path()))
-    for name, path in configuration_paths:
-        platform = spack.platforms.host().name
-        platform_scope = spack.config.ConfigScope(
-            "/".join([name, platform]), os.path.join(path, platform)
-        )
-        generic_scope = spack.config.ConfigScope(name, path)
-        config_scopes.extend([generic_scope, platform_scope])
-        msg = "[BOOTSTRAP CONFIG SCOPE] name={0}, path={1}"
-        tty.debug(msg.format(generic_scope.name, generic_scope.path))
-        tty.debug(msg.format(platform_scope.name, platform_scope.path))
-    return config_scopes
-
-
-def _add_compilers_if_missing():
-    arch = spack.spec.ArchSpec.frontend_arch()
-    if not spack.compilers.compilers_for_arch(arch):
-        new_compilers = spack.compilers.find_new_compilers()
-        if new_compilers:
-            spack.compilers.add_compilers_to_config(new_compilers, init_config=False)
-
-
 def _add_externals_if_missing():
     search_list = [
         # clingo
@@ -656,92 +606,9 @@ def _add_externals_if_missing():
     spack.detection.update_configuration(detected_packages, scope="bootstrap")
 
 
-#: Reference counter for the bootstrapping configuration context manager
-_REF_COUNT = 0
-
-
 def is_bootstrapping():
     global _REF_COUNT
     return _REF_COUNT > 0
-
-
-@contextlib.contextmanager
-def ensure_bootstrap_configuration():
-    # The context manager is reference counted to ensure we don't swap multiple
-    # times if there's nested use of it in the stack. One compelling use case
-    # is bootstrapping patchelf during the bootstrap of clingo.
-    global _REF_COUNT
-    already_swapped = bool(_REF_COUNT)
-    _REF_COUNT += 1
-    try:
-        if already_swapped:
-            yield
-        else:
-            with _ensure_bootstrap_configuration():
-                yield
-    finally:
-        _REF_COUNT -= 1
-
-
-@contextlib.contextmanager
-def _ensure_bootstrap_configuration():
-    bootstrap_store_path = store_path()
-    user_configuration = _read_and_sanitize_configuration()
-    with spack.environment.no_active_environment():
-        with spack.platforms.prevent_cray_detection():
-            with spack.platforms.use_platform(spack.platforms.real_host()):
-                with spack.repo.use_repositories(spack.paths.packages_path):
-                    with spack.store.use_store(bootstrap_store_path):
-                        # Default configuration scopes excluding command line
-                        # and builtin but accounting for platform specific scopes
-                        config_scopes = _bootstrap_config_scopes()
-                        with spack.config.use_configuration(*config_scopes):
-                            # We may need to compile code from sources, so ensure we
-                            # have compilers for the current platform
-                            _add_compilers_if_missing()
-                            spack.config.set("bootstrap", user_configuration["bootstrap"])
-                            spack.config.set("config", user_configuration["config"])
-                            with spack.modules.disable_modules():
-                                with spack_python_interpreter():
-                                    yield
-
-
-def _read_and_sanitize_configuration():
-    """Read the user configuration that needs to be reused for bootstrapping
-    and remove the entries that should not be copied over.
-    """
-    # Read the "config" section but pop the install tree (the entry will not be
-    # considered due to the use_store context manager, so it will be confusing
-    # to have it in the configuration).
-    config_yaml = spack.config.get("config")
-    config_yaml.pop("install_tree", None)
-    user_configuration = {"bootstrap": spack.config.get("bootstrap"), "config": config_yaml}
-    return user_configuration
-
-
-def store_path():
-    """Path to the store used for bootstrapped software"""
-    enabled = spack.config.get("bootstrap:enable", True)
-    if not enabled:
-        msg = "bootstrapping is currently disabled. " 'Use "spack bootstrap enable" to enable it'
-        raise RuntimeError(msg)
-
-    return _store_path()
-
-
-def _root_path():
-    """Root of all the bootstrap related folders"""
-    return spack.config.get("bootstrap:root", spack.paths.default_user_bootstrap_path)
-
-
-def _store_path():
-    bootstrap_root_path = _root_path()
-    return spack.util.path.canonicalize_path(os.path.join(bootstrap_root_path, "store"))
-
-
-def _config_path():
-    bootstrap_root_path = _root_path()
-    return spack.util.path.canonicalize_path(os.path.join(bootstrap_root_path, "config"))
 
 
 def _root_spec(spec_str):
